// backend/src/buyback_catalog/types/faqTypes.ts
export interface TFAQCreate {
  question: string;
  answer: string;
  order_no?: number;
  is_published?: boolean;
  shop_id: number;
  tenant_id: number;
}

// Update FAQ input type
export interface TFAQUpdate {
  question?: string;
  answer?: string;
  order_no?: number;
  is_published?: boolean;
}

// FAQ with translations create type
export interface TFAQWithTranslationsCreate {
  question: string;
  answer: string;
  order_no?: number;
  is_published?: boolean;
  shop_id: number;
  tenant_id: number;
  translations: {
    language_id: number;
    question: string;
    answer: string;
  }[];
}

// FAQ with translations update type  
export interface TFAQWithTranslationsUpdate {
  question?: string;
  answer?: string;
  order_no?: number;
  is_published?: boolean;
  translations?: {
    language_id: number;
    question: string;
    answer: string;
  }[];
}

// FAQ Translation create type
export interface TFAQTranslationCreate {
  question: string;
  answer: string;
}

// FAQ Translation update type
export interface TFAQTranslationUpdate {
  question?: string;
  answer?: string;
}